feat: Payer e Payment adicionados para fazer a comunicação com o mercado pago

+++ src/main/java/com/gabriel_henrique/regua_barbier/domain/servico/ServicoDTO.java
@@ -3,9 +3,11 @@ package com.gabriel_henrique.regua_barbier.domain.servico;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record ServicoDTO(
         @NotBlank String nome,
-        @NotNull Float preco,
+        @NotNull BigDecimal preco,
         @NotNull Integer duracaoMinutos
 ) {
 }
